passing cortisol decades function to separare files
