+++ Output/female/average_cortisol_female.csv
@@ -1,0 +1,5 @@
+0,1.5731766742184359,2.32,0.49317913481413506,0.5178904205577709
+1,1.5797276037151744,2.3296608071304776,0.49523280227201,0.520046989293851
+1,1.5797276037151744,2.3296608071304776,0.49523280227201,0.520046989293851
+1,1.5797276037151744,2.3296608071304776,0.49523280227201,0.520046989293851
+1,1.5797276037151744,2.3296608071304776,0.49523280227201,0.520046989293851
